refactor v.1
