+++ src/main/java/com/board_of_ads/controllers/rest/UserRestController.java
@@ -3,6 +3,7 @@ package com.board_of_ads.controllers.rest;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
+
 import java.security.Principal;
 
 @RestController
