+++ src/main/resources/application.yml
@@ -6,7 +6,7 @@ spring:
   datasource:
     driver-class-name: org.postgresql.Driver
     url: jdbc:postgresql://localhost:5432/board_of_ads
-    username: root
+    username: postgres
     password: root
     initialization-mode: always
